NFDiv-2697 Rename switch.to.sole.content.*
