+++ src/main/steps/common/switch-to-sole-content.utils.ts
